fix float16 training

+++ util/utils.py
@@ -167,8 +167,8 @@ def pose_err(est_pose, gt_pose):
     posit_err = torch.norm(est_pose[:, 0:3] - gt_pose[:, 0:3], dim=1)
     est_pose_q = F.normalize(est_pose[:, 3:], p=2, dim=1)
     gt_pose_q = F.normalize(gt_pose[:, 3:], p=2, dim=1)
-    inner_prod = torch.bmm(est_pose_q.view(est_pose_q.shape[0], 1, est_pose_q.shape[1]),
-                           gt_pose_q.view(gt_pose_q.shape[0], gt_pose_q.shape[1], 1))
+    inner_prod = torch.bmm(est_pose_q.float().view(est_pose_q.shape[0], 1, est_pose_q.shape[1]),
+                           gt_pose_q.float().view(gt_pose_q.shape[0], gt_pose_q.shape[1], 1))
     orient_err = 2 * torch.acos(torch.abs(inner_prod)) * 180 / np.pi
     return posit_err, orient_err
